LookupByCorrelationID returns a list of jobs
Also, List allows filtering by correlation identifier now. The reason
for this change is that the correlation identifier might not be unique
in the application using jobqueue. However, that means the
LookupByCorrelationID only returns one result of the jobs regarding the
correlated entity. That, again, might lead to wrong decisions in the
application, depending on the single entity returned.